+++ R/print.hdi.R
@@ -38,13 +38,13 @@ print.ci <- function(x, digits = 2, ...) {
 
   if (length(ci) == 1) {
     xsub <- .remove_column(x, c("CI", "CI_low", "CI_high"))
-    colnames(xsub)[ncol(xsub)] <- sprintf("%i%% %s", ci, ci_string)
+    colnames(xsub)[ncol(xsub)] <- sprintf("%.5g%% %s", ci, ci_string)
     print_data_frame(xsub, digits = digits)
   } else {
     for (i in ci) {
       xsub <- x[x$CI == i, -which(colnames(x) == "CI"), drop = FALSE]
       xsub <- .remove_column(xsub, c("CI", "CI_low", "CI_high"))
-      colnames(xsub)[ncol(xsub)] <- sprintf("%i%% %s", i, ci_string)
+      colnames(xsub)[ncol(xsub)] <- sprintf("%.5g%% %s", i, ci_string)
       print_data_frame(xsub, digits = digits)
       cat("\n")
     }