+++ tests/testthat/test-describe_posterior.R
@@ -25,25 +25,26 @@ test_that("describe_posterior", {
   library(rstanarm)
   x <- insight::download_model("stanreg_lm_1")
   rez <- describe_posterior(x, estimate = "all", dispersion = TRUE, test = "all")
-  testthat::expect_equal(dim(rez), c(2, 18))
+  testthat::expect_equal(dim(rez), c(2, 21))
   rez <- describe_posterior(x, estimate = "all", dispersion = TRUE, test = "all", ci = c(0.8, 0.9))
-  testthat::expect_equal(dim(rez), c(4, 18))
-  rez <- describe_posterior(x, estimate = NULL, dispersion = TRUE, test = NULL, ci_method = "quantile")
-  testthat::expect_equal(dim(rez), c(2, 6))
+  testthat::expect_equal(dim(rez), c(4, 21))
+  rez <- describe_posterior(x, estimate = NULL, dispersion = TRUE, test = NULL, ci_method = "quantile", diagnostic = NULL, priors=FALSE)
+  testthat::expect_equal(dim(rez), c(2, 4))
 
   # Brms
   library(brms)
   x <- insight::download_model("brms_mixed_1")
-  # rez <- describe_posterior(x, estimate = "all", dispersion = TRUE, test = "all")
+  # rez <- describe_posterior(x, estimate = "all", dispersion = TRUE, test = "all")  # doenst work because of BFs
   # testthat::expect_equal(dim(rez), c(4, 16))
-  # rez <- describe_posterior(x, estimate = "all", dispersion = TRUE, test = "all", ci = c(0.8, 0.9))
-  # testthat::expect_equal(dim(rez), c(8, 16))
-  # rez <- describe_posterior(x, estimate = NULL, dispersion = TRUE, test = NULL, ci_method="quantile")
-  # testthat::expect_equal(dim(rez), c(4, 4))
+  rez <- describe_posterior(x, estimate = "all", dispersion = TRUE, ci = c(0.8, 0.9))
+  testthat::expect_equal(dim(rez), c(4, 16))
+  rez <- describe_posterior(x, estimate = NULL, dispersion = TRUE, test = NULL, ci_method = "quantile", diagnostic = NULL)
+  testthat::expect_equal(dim(rez), c(2, 4))
+
 
   # BayesFactor
-  library(BayesFactor)
-  x <- ttestBF(x = rnorm(100, 1, 1))
+  # library(BayesFactor)
+  # x <- BayesFactor::ttestBF(x = rnorm(100, 1, 1))
   # rez <- describe_posterior(x, estimate = "all", dispersion = TRUE, test = "all")
   # testthat::expect_equal(dim(rez), c(4, 16))
   # rez <- describe_posterior(x, estimate = "all", dispersion = TRUE, test = "all", ci = c(0.8, 0.9))