+++ NEWS.md
@@ -15,6 +15,7 @@
 
 ## Minor changes
 
+- Improved documentation
 - `p_direction()`: improved printing
 - Improved testing
 - `rope()` for model-objects now returns the HDI values for all parameters as attribute in a consistent way